requirements.txt
kafka-python
boto3
pandas
pyarrow
python-dateutil
requests
